Qt/patches: move game list loop to patch manager
This deduplicates some code
