+++ rpcs3/rpcs3.vcxproj.filters
@@ -1164,6 +1164,9 @@
     <ClCompile Include="rpcs3qt\permissions.cpp">
       <Filter>Gui\utils</Filter>
     </ClCompile>
+    <ClCompile Include="rpcs3qt\gui_game_info.cpp">
+      <Filter>Gui\game list</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Input\ds4_pad_handler.h">
@@ -1369,6 +1372,9 @@
     </ClInclude>
     <ClInclude Include="rpcs3qt\permissions.h">
       <Filter>Gui\utils</Filter>
+    </ClInclude>
+    <ClInclude Include="rpcs3qt\gui_game_info.h">
+      <Filter>Gui\game list</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
