+++ rpcs3/rpcs3qt/gui_game_info.cpp
@@ -1,0 +1,14 @@
+#include "stdafx.h"
+#include "gui_game_info.h"
+#include "localized.h"
+
+std::string gui_game_info::GetGameVersion() const
+{
+	if (info.app_ver == Localized().category.unknown.toStdString())
+	{
+		// Fall back to Disc/Pkg Revision
+		return info.version;
+	}
+
+	return info.app_ver;
+}
